refactor: apis/user 폴더구조 변경
