+++ src/components/layouts/root/navbar/login-button.tsx
@@ -3,7 +3,7 @@
 import useIsMobile from '@/lib/hooks/useIsMobile';
 import { useCurrentUserInfoStore } from '@/lib/store/useCurrentUserInfoStore';
 import { useIsLoginModalOpenStore } from '@/lib/store/useIsLoginModalOpenStore';
-import { getUserInfo } from '@/services/apis/user';
+import { getUserInfo } from '@/services/apis/user/user.api';
 import clsx from 'clsx';
 import Image from 'next/image';
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
