+++ src/services/apis/shorts/shorts.type.ts
@@ -1,5 +1,5 @@
 import { SuccessResponse } from '@/services/config/dto';
-import { UserDto } from '../user/dto';
+import { UserDto } from '../user/user.type';
 
 // enum
 export type HalftimeSortType = 'CREATED_DESC' | 'POPULAR' | 'CREATED_ASC';
